refactor: remove UploadControls component and integrate its functionality into StatsSummary, adding ShowMeHowDialog for user guidance
